+++ app/Filament/Resources/RelationManagers/AbstractControllersRelationManager.php
@@ -5,7 +5,6 @@ namespace App\Filament\Resources\RelationManagers;
 use App\Filament\Resources\TranslatesStrings;
 use App\Helpers\Controller\FrequencyFormatter;
 use App\Models\Controller\ControllerPosition;
-use App\Models\Controller\HasControllerHierarchy;
 use App\Services\ControllerPositionHierarchyService;
 use Filament\Forms;
 use Filament\Resources\RelationManagers\RelationManager;
@@ -48,18 +47,22 @@ abstract class AbstractControllersRelationManager extends RelationManager
                                 $livewire->getOwnerRecord()
                                     ->controllers
                                     ->mapWithKeys(
-                                        fn (ControllerPosition $controller) => [$controller->id => $controller->callsign]
+                                        fn (ControllerPosition $controller) =>
+                                            [$controller->id => $controller->callsign]
                                     )
                             ),
                     ])
                     ->using(function (AbstractControllersRelationManager $livewire, $data) {
-                        self::doUpdate(fn () => ControllerPositionHierarchyService::insertPositionIntoHierarchy(
-                            $livewire->getOwnerRecord(),
-                            ControllerPosition::findOrFail($data['recordId']),
-                            after: isset($data['insert_after'])
-                                ? ControllerPosition::findOrFail($data['insert_after'])
-                                : null
-                        ), $livewire->getOwnerRecord());
+                        self::doUpdate(
+                            fn () => ControllerPositionHierarchyService::insertPositionIntoHierarchy(
+                                $livewire->getOwnerRecord(),
+                                ControllerPosition::findOrFail($data['recordId']),
+                                after: isset($data['insert_after'])
+                                    ? ControllerPosition::findOrFail($data['insert_after'])
+                                    : null
+                            ),
+                            $livewire->getOwnerRecord()
+                        );
                     })
                     ->disableAttachAnother()
                     ->label(self::translateTablePath('attach_action.label'))
@@ -69,32 +72,41 @@ abstract class AbstractControllersRelationManager extends RelationManager
             ->actions([
                 Tables\Actions\Action::make('moveUp')
                     ->action(function (ControllerPosition $record) {
-                        self::doUpdate(fn () => ControllerPositionHierarchyService::moveControllerInHierarchy(
-                            $record->pivot->pivotParent,
-                            $record,
-                            true
-                        ), $record->pivot->pivotParent);
+                        self::doUpdate(
+                            fn () => ControllerPositionHierarchyService::moveControllerInHierarchy(
+                                $record->pivot->pivotParent,
+                                $record,
+                                true
+                            ),
+                            $record->pivot->pivotParent
+                        );
                     })
                     ->label(self::translateTablePath('move_up_action.label'))
                     ->icon('heroicon-o-arrow-up')
                     ->authorize(fn (AbstractControllersRelationManager $livewire) => $livewire->can('moveUp')),
                 Tables\Actions\Action::make('moveDown')
                     ->action(function (ControllerPosition $record) {
-                        self::doUpdate(fn () => ControllerPositionHierarchyService::moveControllerInHierarchy(
-                            $record->pivot->pivotParent,
-                            $record,
-                            false
-                        ), $record->pivot->pivotParent);
+                        self::doUpdate(
+                            fn () => ControllerPositionHierarchyService::moveControllerInHierarchy(
+                                $record->pivot->pivotParent,
+                                $record,
+                                false
+                            ),
+                            $record->pivot->pivotParent
+                        );
                     })
                     ->label(self::translateTablePath('move_down_action.label'))
                     ->icon('heroicon-o-arrow-down')
                     ->authorize(fn (AbstractControllersRelationManager $livewire) => $livewire->can('moveUp')),
                 Tables\Actions\DetachAction::make()
                     ->using(function (ControllerPosition $record) {
-                        self::doUpdate(fn () => ControllerPositionHierarchyService::removeFromHierarchy(
-                            $record->pivot->pivotParent,
-                            $record
-                        ), $record->pivot->pivotParent);
+                        self::doUpdate(
+                            fn () => ControllerPositionHierarchyService::removeFromHierarchy(
+                                $record->pivot->pivotParent,
+                                $record
+                            ),
+                            $record->pivot->pivotParent
+                        );
                     })->label(self::translateTablePath('detach_action.label')),
             ]);
     }