+++ app/Services/AircraftService.php
@@ -12,7 +12,9 @@ class AircraftService
         return Aircraft::with('wakeCategories')->get()->map(fn (Aircraft $aircraft) => [
             'id' => $aircraft->id,
             'icao_code' => $aircraft->code,
-            'wake_categories' => $aircraft->wakeCategories->map(fn (WakeCategory $category) => $category->id)->toArray(),
+            'wake_categories' => $aircraft->wakeCategories->map(
+                fn (WakeCategory $category) => $category->id
+            )->toArray(),
         ])->toArray();
     }
 }