+++ config/srd.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'download_url' => env('SRD_DOWNLOAD_URL'),
+];