+++ lang/en/srd/form.php
@@ -1,0 +1,19 @@
+<?php
+
+return [
+    'origin' => [
+        'label' => 'Origin',
+    ],
+    'destination' => [
+        'label' => 'Destination',
+    ],
+    'minimum_level' => [
+        'label' => 'Minimum Level',
+    ],
+    'maximum_level' => [
+        'label' => 'Maximum Level',
+    ],
+    'route_segment' => [
+        'label' => 'Route Segment',
+    ],
+];